Fixed remove product function in cart to process long values, not int(didnt work correct before)

+++ EshopSpringBoot/src/main/java/by/tms/eshopspringboot/controller/CartController.java
@@ -42,7 +42,7 @@ public class CartController {
     }
 
     @DeleteMapping
-    public void removeProductFromCart(@SessionAttribute Map<Integer, Integer> cartProductsMap, @RequestParam(PRODUCT_ID) int productId) {
+    public void removeProductFromCart(@SessionAttribute Map<Long, Integer> cartProductsMap, @RequestParam(PRODUCT_ID) Long productId) {
         if (cartProductsMap.get(productId) > 1) {
             cartProductsMap.compute(productId, (key, value) -> value - 1);
         } else {
